requirements.txt
streamlit>=1.30
pytest>=7.0
openai>=1.0
google-genai>=1.0
